Implements Antivirus (AV) Configuration Options
* Separates the responsibilities of AV unit tests
* Provides user options for which AV scanner to use
* Provides user options for max_file_size and max_scan_size
* Ensures PREMIS events are not created for unscanned, and ostensibly unscanned files
* Fixes #779
* Resolves #819
* Additional code review changes
